Added code to service classes

+++ src/main/java/org/keyin/Airport/Airport.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 public class Airport {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected int id;
+    protected Long id;
     protected String name;
 
     @ManyToMany
